fix: move task repo to repos crate
