+++ crates/repos/src/postgres.rs
@@ -1,13 +1,13 @@
 use crate::{
-    entity::{Comment, Episode, Podcast, Script},
+    entity::{Comment, Episode, Podcast, Script, Task, TaskStatus},
     error::Error,
     repo::{
         CommentId, CommentRepo, EpisodeId, EpisodeRepo, PodcastId, PodcastRepo, ScriptId,
-        ScriptRepo,
+        ScriptRepo, TaskId, TaskRepo,
     },
 };
 use async_trait::async_trait;
-use chrono::Local;
+use chrono::{Local, Utc};
 use sqlx::{PgPool, Pool, Postgres};
 use std::sync::LazyLock;
 use uuid::Uuid;
@@ -232,6 +232,71 @@ impl ScriptRepo for DummyScriptRepo {
     }
 
     async fn update(&self, _script: &Script) -> anyhow::Result<()> {
+        Ok(())
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct PostgresTaskRepo {
+    pool: Pool<Postgres>,
+}
+
+impl PostgresTaskRepo {
+    pub fn new() -> Self {
+        let pool = PG_POOL.clone();
+        Self { pool }
+    }
+}
+
+#[async_trait]
+impl TaskRepo for PostgresTaskRepo {
+    async fn pop(&self) -> anyhow::Result<Option<Task>> {
+        let task = sqlx::query_as!(
+            Task,
+            r#"select id, status as "status!: TaskStatus", args, execute_after, executed_at from tasks where status = $1 and $2 < execute_after order by id limit 1"#,
+            TaskStatus::Pending as TaskStatus,
+            Utc::now(),
+        )
+        .fetch_optional(&self.pool)
+        .await?;
+        Ok(task)
+    }
+
+    async fn create(&self, task: &Task) -> anyhow::Result<()> {
+        sqlx::query_as!(
+            Task,
+            "insert into tasks (id, status, args, execute_after, executed_at) values ($1, $2, $3, $4, $5)",
+            task.id,
+            &task.status as &TaskStatus,
+            task.args,
+            task.execute_after,
+            task.executed_at,
+        )
+        .execute(&self.pool)
+        .await?;
+        Ok(())
+    }
+
+    async fn update(&self, task: &Task) -> anyhow::Result<()> {
+        sqlx::query_as!(
+            Task,
+            "update tasks set status = $2, args = $3, execute_after = $4, executed_at = $5 where id = $1",
+            task.id,
+            &task.status as &TaskStatus,
+            task.args,
+            task.execute_after,
+            task.executed_at,
+        )
+        .execute(&self.pool)
+        .await?;
+        Ok(())
+    }
+
+    async fn delete(&self, id: &TaskId) -> anyhow::Result<()> {
+        sqlx::query("delete from tasks where id = $1")
+            .bind(id.0)
+            .fetch_one(&self.pool)
+            .await?;
         Ok(())
     }
 }
